+++ test/snippets/snippets.cpp
@@ -129,192 +129,7 @@ void run_coro(boost::asio::any_io_executor ex, std::function<boost::asio::awaita
 }
 #endif
 
-std::int64_t get_employee_id() { return 42; }
 std::string get_company_id() { return "HGS"; }
-
-// Describe types
-
-//[describe_post
-// We can use a plain struct with ints and strings to describe our rows.
-// This must be placed at the namespace level
-struct post
-{
-    int id;
-    std::string title;
-    std::string body;
-};
-
-// We must use Boost.Describe to add reflection capabilities to post.
-// We must list all the fields that should be populated by Boost.MySQL
-BOOST_DESCRIBE_STRUCT(post, (), (id, title, body))
-//]
-
-//[describe_stored_procedures
-// Describes the first resultset
-struct company
-{
-    std::string id;
-    std::string name;
-    std::string tax_id;
-};
-BOOST_DESCRIBE_STRUCT(company, (), (id, name, tax_id))
-
-// Describes the second resultset
-struct employee
-{
-    std::string first_name;
-    std::string last_name;
-    boost::optional<std::uint64_t> salary;
-};
-BOOST_DESCRIBE_STRUCT(employee, (), (first_name, last_name, salary))
-
-// The last resultset will always be empty.
-// We can use an empty tuple to represent it.
-using empty = std::tuple<>;
-//]
-
-void section_multi_resultset(tcp_ssl_connection& conn)
-{
-    {
-        //[multi_resultset_call_dynamic
-
-        // We're using the dynamic interface. results can stored multiple resultsets
-        results result;
-
-        // The procedure parameter, employe_id, will likely be obtained from an untrusted source,
-        // so we will use a prepared statement
-        statement get_employee_stmt = conn.prepare_statement("CALL get_employees(?)");
-
-        // Obtain the parameters required to call the statement, e.g. from a file or HTTP message
-        std::int64_t employee_id = get_employee_id();
-
-        // Call the statement
-        conn.execute(get_employee_stmt.bind(employee_id), result);
-
-        // results can be used as a random-access collection of resultsets.
-        // result.at(0).rows() returns the matched companies, if any
-        rows_view matched_company = result.at(0).rows();
-
-        // We can do the same to access the matched employees
-        rows_view matched_employees = result.at(1).rows();
-
-        // Use matched_company and matched_employees as required
-        //]
-
-        boost::ignore_unused(matched_company);
-        boost::ignore_unused(matched_employees);
-    }
-#ifdef BOOST_MYSQL_CXX14
-    {
-        //[multi_resultset_call_static
-        // We must list all the resultset types the operation returns as template arguments
-        static_results<company, employee, empty> result;
-        conn.execute("CALL get_employees('HGS')", result);
-
-        // We can use rows<0>() to access the rows for the first resultset
-        if (result.rows<0>().empty())
-        {
-            std::cout << "Company not found" << std::endl;
-        }
-        else
-        {
-            const company& comp = result.rows<0>()[0];
-            std::cout << "Company name: " << comp.name << ", tax_id: " << comp.tax_id << std::endl;
-        }
-
-        // rows<1>() will return the rows for the second resultset
-        for (const employee& emp : result.rows<1>())
-        {
-            std::cout << "Employee " << emp.first_name << " " << emp.last_name << std::endl;
-        }
-        //]
-    }
-#endif
-    {
-        //[multi_resultset_out_params
-        // To retrieve output parameters, you must use prepared statements. Text queries don't support this
-        // We specify placeholders for both IN and OUT parameters
-        statement stmt = conn.prepare_statement("CALL create_employee(?, ?, ?, ?)");
-
-        // When executing the statement, we provide an actual value for the IN parameters,
-        // and a dummy value for the OUT parameter. This value will be ignored, but it's required by the
-        // protocol
-        results result;
-        conn.execute(stmt.bind("HGS", "John", "Doe", nullptr), result);
-
-        // Retrieve output parameters. This row_view has an element per
-        // OUT or INOUT parameter that used a ? placeholder
-        row_view output_params = result.out_params();
-        std::int64_t new_employee_id = output_params.at(0).as_int64();
-        //]
-
-        boost::ignore_unused(new_employee_id);
-    }
-}
-
-void section_multi_resultset_multi_queries()
-{
-    boost::asio::io_context ctx;
-    boost::asio::ssl::context ssl_ctx(boost::asio::ssl::context::tls_client);
-    boost::asio::ip::tcp::resolver resolver(ctx.get_executor());
-    boost::mysql::tcp_ssl_connection conn(ctx.get_executor(), ssl_ctx);
-
-    auto endpoint = *resolver.resolve(server_hostname, boost::mysql::default_port_string).begin();
-
-    //[multi_resultset_multi_queries
-    // The username and password to use
-    boost::mysql::handshake_params params(
-        mysql_username,         // username, as a string
-        mysql_password,         // password, as a string
-        "boost_mysql_examples"  // database
-    );
-
-    // Allows running multiple semicolon-separated in a single call.
-    // We must set this before calling connect
-    params.set_multi_queries(true);
-
-    // Connect to the server specifying that we want support for multi-queries
-    conn.connect(endpoint, params);
-
-    // We can now use the multi-query feature.
-    // This will result in three resultsets, one per query.
-    results result;
-    conn.execute(
-        R"(
-            CREATE TEMPORARY TABLE posts (
-                id INT PRIMARY KEY AUTO_INCREMENT,
-                title VARCHAR (256),
-                body TEXT
-            );
-            INSERT INTO posts (title, body) VALUES ('Breaking news', 'Something happened!');
-            SELECT COUNT(*) FROM posts;
-        )",
-        result
-    );
-    //]
-
-    //[multi_resultset_results_as_collection
-    // result is actually a random-access collection of resultsets.
-    // The INSERT is the 2nd query, so we can access its resultset like this:
-    boost::mysql::resultset_view insert_result = result.at(1);
-
-    // A resultset has metadata, rows, and additional data, like the last insert ID:
-    std::int64_t post_id = insert_result.last_insert_id();
-
-    // The SELECT result is the third one, so we can access it like this:
-    boost::mysql::resultset_view select_result = result.at(2);
-
-    // select_result is a view that points into result.
-    // We can take ownership of it using the resultse class:
-    boost::mysql::resultset owning_select_result(select_result);  // valid even after result is destroyed
-
-    // We can access rows of resultset objects as usual:
-    std::int64_t num_posts = owning_select_result.rows().at(0).at(0).as_int64();
-    //]
-
-    boost::ignore_unused(post_id);
-    boost::ignore_unused(num_posts);
-}
 
 void section_multi_function(tcp_ssl_connection& conn)
 {
